move appUpdater and diskLRU and sentry to service
