+++ src/services/appUpdater.ts
@@ -2,7 +2,7 @@ import { BrowserWindow, dialog, Tray } from 'electron';
 import { autoUpdater } from 'electron-updater';
 import log from 'electron-log';
 import { setIsAppQuitting, setIsUpdateAvailable } from '../main';
-import { buildContextMenu } from './menu';
+import { buildContextMenu } from '../utils/menu';
 
 class AppUpdater {
     constructor() {
