+++ src/services/sentry.ts
@@ -1,14 +1,14 @@
 import * as Sentry from '@sentry/electron/dist/main';
+import { keysStore } from '../stores/keys.store';
 
-import { keysStore } from '../services/store';
-import { isDev } from './common';
+import { isDev } from '../utils/common';
 
 const SENTRY_DSN = 'https://e9268b784d1042a7a116f53c58ad2165@sentry.ente.io/5';
 
 // eslint-disable-next-line @typescript-eslint/no-var-requires
 const version = require('../../package.json').version;
 
-function initSentry(): void {
+export function initSentry(): void {
     Sentry.init({
         dsn: SENTRY_DSN,
         release: version,
@@ -46,6 +46,15 @@ function errorWithContext(originalError: Error, context: string) {
     return errorWithContext;
 }
 
+function getUserAnonymizedID() {
+    let anonymizeUserID = keysStore.get('AnonymizeUserID')?.id;
+    if (!anonymizeUserID) {
+        anonymizeUserID = makeID(6);
+        keysStore.set('AnonymizeUserID', { id: anonymizeUserID });
+    }
+    return anonymizeUserID;
+}
+
 function makeID(length: number) {
     let result = '';
     const characters =
@@ -58,14 +67,3 @@ function makeID(length: number) {
     }
     return result;
 }
-
-function getUserAnonymizedID() {
-    let anonymizeUserID = keysStore.get('AnonymizeUserID')?.id;
-    if (!anonymizeUserID) {
-        anonymizeUserID = makeID(6);
-        keysStore.set('AnonymizeUserID', { id: anonymizeUserID });
-    }
-    return anonymizeUserID;
-}
-
-export default initSentry;
